Starting work refactoring axes and scales
