+++ assets/src/scripts/components/dailyValueHydrograph/axes.js
@@ -1,0 +1,6 @@
+
+import {createSelector} from 'reselect';
+
+export const getXAxis = createSelector(
+    scale
+);
