+++ assets/src/scripts/components/dailyValueHydrograph/scales.js
@@ -1,0 +1,2 @@
+
+export const getXScale =
