+++ assets/src/scripts/components/hydrograph/axes.js
@@ -4,9 +4,8 @@ import { DateTime } from 'luxon';
 import { createSelector } from 'reselect';
 
 import config from '../../config';
-import {appendXAxis, appendYAxis, appendSecondaryYAxis} from '../../d3-utils/axes';
 import { getCurrentDateRange, getCurrentParmCd } from '../../selectors/time-series-selector';
-import { convertCelsiusToFahrenheit, convertFahrenheitToCelsius, mediaQuery, wrap, deltaDays } from '../../utils';
+import { convertCelsiusToFahrenheit, convertFahrenheitToCelsius, mediaQuery, deltaDays } from '../../utils';
 
 import { getYTickDetails } from './domain';
 import {getLayout} from './layout';
@@ -204,71 +203,3 @@ export const getAxes = memoize(kind => createSelector(
         };
     }
 ));
-
-/*
- * Add X Axis to svg or group elem given the xAxis and layout
- * @param {Object} elem - svg or g D3 selection
- * @param {Object}
- *      @prop {Object} xAxis - D3 axis object
- *      @prop {Object} layout - contains properties for width, height, and margin for enclosing svg.
- */
-export const appendXAxis = function(elem, {xAxis, layout}) {
-    const xLoc = {
-        x: 0,
-        y: layout.height - (layout.margin.top + layout.margin.bottom)
-    };
-    elem.selectAll('.x-axis').remove();
-    elem.append('g')
-        .attr('class', 'x-axis')
-        .attr('transform', `translate(${xLoc.x}, ${xLoc.y})`)
-        .call(xAxis);
-};
-
-/**
- * Add x and y axes to the given svg node.
- */
-export const appendAxes = function(elem, {xAxis, yAxis, secondaryYAxis, layout, yTitle, secondaryYTitle}) {
-
-    const yLoc = {x: 0, y: 0};
-    const yLabelLoc = {
-        x: layout.height / -2 + layout.margin.top,
-        y: -1 * layout.margin.left + 12
-    };
-
-    appendXAxis(elem, {xAxis, layout});
-
-    // Remove existing axes before adding the new ones.
-    elem.selectAll('.y-axis').remove();
-
-    // Add y-axis and a text label
-    elem.append('g')
-        .attr('class', 'y-axis')
-        .attr('transform', `translate(${yLoc.x}, ${yLoc.y})`)
-        .call(yAxis)
-        .append('text')
-            .attr('class', 'y-axis-label')
-            .attr('transform', 'rotate(-90)')
-            .attr('x', yLabelLoc.x)
-            .attr('y', yLabelLoc.y)
-            .text(yTitle)
-                .call(wrap, layout.height - (layout.margin.top + layout.margin.bottom));
-
-    if (secondaryYAxis !== null && secondaryYTitle !== null) {
-        const maxXScaleRange = xAxis.scale().range()[1];
-        const secondaryYLabelLoc = {
-            x: layout.height / -2 + layout.margin.top,
-            y: (layout.width - maxXScaleRange) * 1.5
-        };
-        elem.append('g')
-            .attr('class', 'y-axis')
-            .attr('transform', `translate(${maxXScaleRange}, ${yLoc.y})`)
-            .call(secondaryYAxis)
-            .append('text')
-                .attr('class', 'y-axis-label')
-                .attr('transform', 'rotate(-90)')
-                .attr('x', secondaryYLabelLoc.x )
-                .attr('y', secondaryYLabelLoc.y )
-                .text(secondaryYTitle)
-                    .call(wrap, layout.height - (layout.margin.top + layout.margin.bottom));
-    }
-};
