+++ assets/src/scripts/d3-rendering/accessibility.js
