+++ assets/src/scripts/d3-rendering/accessibility.spec.js
