+++ assets/src/scripts/d3-rendering/loading-indicator.spec.js
